requirements.txt
mistralai>=0.1.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
moviepy>=1.0.3
flask>=3.0.0
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.5
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
